requirements.txt
Django==1.9.1
django-crispy-forms==1.6.0
django-filter==0.12.0
django-guardian==1.4.1
djangorestframework==3.3.2
Markdown==2.6.5
six==1.10.0
wheel==0.24.0
